api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tidb-oracle-api"
version = "0.1.0"
description = "TiDB Oracle internal GTM copilot backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.18",
  "pgvector>=0.3.2",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.3.4",
  "python-dotenv>=1.0.1",
  "typer>=0.12.3",
  "celery>=5.4.0",
  "redis>=5.0.7",
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "python-dateutil>=2.9.0",
  "email-validator>=2.2.0",
  "jinja2>=3.1.4"
  ,
  "google-api-python-client>=2.151.0",
  "google-auth>=2.34.0",
  "google-auth-oauthlib>=1.2.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "anyio>=4.4.0"
]

[project.scripts]
kb = "app.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
